Refactor simple job spec expansion
